Drop the term sliver and add gap
