+++ geoplanar/__init__.py
@@ -6,8 +6,8 @@ A module for handling planar enforcement for polygon geoseries.
 """
 
 from geoplanar.overlap import trim_overlaps, is_overlapping, overlaps
-from geoplanar.hole import (holes, fill_holes, add_interiors,
-                            missing_interiors)
+from geoplanar.gap import (gaps, fill_gaps)
+from geoplanar.hole import (add_interiors, missing_interiors)
 from geoplanar.planar import (planar_enforce, is_planar_enforced,
                               fix_npe_edges, insert_intersections,
                               non_planar_edges, check_validity,
