+++ geoplanar/hole.py
@@ -6,127 +6,6 @@ from shapely.geometry import box
 import shapely
 
 
-def holes(gdf):
-    """Find holes in a geodataframe.
-
-    A hole is also known as a sliver which contains a set of points that:
-
-    - are not contained by any of the geometries in the geoseries
-    - are contained in the convex hull of the unary_union of the geoseries
-
-    Parameters
-    ----------
-
-    gdf :  GeoDataFrame with polygon (multipolygon) GeoSeries
-
-
-    Returns
-    -------
-
-    _holes : GeoDataFrame with hole polygons
-
-    Examples
-    --------
-    >>> p1 = box(0,0,10,10)
-    >>> p2 = Polygon([(10,10), (12,8), (10,6), (12,4), (10,2), (20,5)])
-    >>> gdf = geopandas.GeoDataFrame(geometry=[p1,p2])
-    >>> h = geoplanar.holes(gdf)
-    >>> h.area
-    array([4., 4.])
-    """
-
-    # get box
-    b = box(*gdf.total_bounds)
-
-    # form unary union
-    u = gdf.unary_union
-
-    # diff of b and u
-    dbu = geopandas.GeoDataFrame(geometry=[b.difference(u)])
-
-    # explode
-    _holes = dbu.explode()
-
-
-    _holes.reset_index(inplace=True)
-
-    gaps = []
-    for idx, row in _holes.iterrows():
-        its = b.exterior.intersection(row.geometry)
-        if not isinstance(its, shapely.geometry.multilinestring.MultiLineString):
-            gaps.append(idx)
-
-    return _holes.iloc[gaps]
-
-
-
-def fill_holes(gdf, largest=True, inplace=False):
-    """Fill any holes in a geodataframe.
-
-    A hole is also known as a sliver which contains a set of points that:
-
-    - are not contained by any of the geometries in the geoseries
-    - are contained in the convex hull of the unary_union of the geoseries
-
-    Parameters
-    ----------
-
-    gdf :  GeoDataFrame with polygon (multipolygon) GeoSeries
-
-
-    largest: boolean (Default: True)
-          Merge each hole with its largest (True), or smallest (False) touching neighbor
-
-    inplace: boolean (default: False)
-          Change the geoseries of current dataframe
-
-
-    Returns
-    -------
-
-    _holes : GeoDataFrame with hole polygons
-
-    Examples
-    --------
-    >>> p1 = box(0,0,10,10)
-    >>> p2 = Polygon([(10,10), (12,8), (10,6), (12,4), (10,2), (20,5)])
-    >>> gdf = geopandas.GeoDataFrame(geometry=[p1,p2])
-    >>> gdf.area
-    0    100.0
-    1     32.0
-    dtype: float64
-    >>> h = geoplanar.holes(gdf)
-    >>> h.area
-    array([4., 4.])
-    >>> gdf1 = geoplanar.fill_holes(gdf)
-    >>> gdf1.area
-    0    108.0
-    1     32.0
-    dtype: float64
-    """
-
-
-    h = holes(gdf)
-    if not inplace:
-        gdf = gdf.copy()
-
-    for index, row in h.iterrows():
-        rdf = geopandas.GeoDataFrame(geometry=[row.geometry])
-        neighbors = geopandas.sjoin(left_df=gdf, right_df=rdf, how='inner',
-                                    op='intersects')
-        if largest:
-            left = neighbors[neighbors.area==neighbors.area.max()]
-        else:
-            left = neighbors[neighbors.area==neighbors.area.min()]
-        tmpdf = pandas.concat([left, rdf]).dissolve()
-        try:
-            geom = tmpdf.loc[0, 'geometry']
-            gdf.geometry[left.index] = geopandas.GeoDataFrame(geometry=[geom]).geometry.values
-        except:
-            print(index, tmpdf.shape)
-
-    return gdf
-
 def missing_interiors(gdf):
     """ Find any missing interiors.
